Restructured vqe example to multiple files.
